Remove directory resolution logs from the server.
This is an unstable branch that will be rebased.

The VolumeDiskData version stamp has been incremented so that will will not run
with existing replicated volumes because it ignores the fact that they have a
resolution log at all and as such will get the volume state out of sync with
the logged state.

This is a work-in-progress trying to aggressively remove all traces of
resolution logs. There is right now no proper migration from the old server RVM
state to the new. Maybe norton-reinit will have to be used to dump RVM state of
the old servers load it back without the resolution logs.

Current status
 - 62 files changed, 83 insertions(+), 8240 deletions(-)
 - everything compiles
 - untested if stripped down servers actually work.

In theory a server without resolution logs will still operate normally for
replicated volumes, except it can only resolve the 'weak-equality' type
directory conflicts and will mark the directory in conflict for anything else.